+++ core/execution/spot/market_test.go
@@ -192,23 +192,7 @@ func newTestMarket(
 	baseAsset := NewAssetStub(base, baseDP)
 	quoteAsset := NewAssetStub(quote, quoteDP)
 
-	market, _ := spot.NewMarket(
-		context.Background(),
-		log,
-		matching.NewDefaultConfig(),
-		fee.NewDefaultConfig(),
-		liquidity.NewDefaultConfig(),
-		collateral,
-		&mkt,
-		ts,
-		broker,
-		as,
-		statevarEngine,
-		mat,
-		baseAsset,
-		quoteAsset,
-		peggedOrderCounterForTest,
-	)
+	market, _ := spot.NewMarket(log, matching.NewDefaultConfig(), fee.NewDefaultConfig(), liquidity.NewDefaultConfig(), collateral, &mkt, ts, broker, as, statevarEngine, mat, baseAsset, quoteAsset, peggedOrderCounterForTest)
 
 	tm := &testMarket{
 		market:           market,