+++ core/types/team.go
@@ -25,17 +25,9 @@ type Team struct {
 }
 
 type Membership struct {
-	PartyID       PartyID
-	JoinedAt      time.Time
-	NumberOfEpoch uint64
-}
-
-func (t *Team) AddReferee(partyID PartyID, joinedAt time.Time) {
-	t.Referees = append(t.Referees, &Membership{
-		PartyID:       partyID,
-		JoinedAt:      joinedAt,
-		NumberOfEpoch: 0,
-	})
+	PartyID        PartyID
+	JoinedAt       time.Time
+	StartedAtEpoch uint64
 }
 
 func (t *Team) RemoveReferee(refereeToRemove PartyID) {